criu: Version 1.5.1
With 1.5.1 we forestall criu crash on soon to be released
4.0 kernel due to uninitialized ss on restore sigframe.

Another thing we "fix" is legalize swrk API and add the
ability for inheriting fds via it. This is required for
libcontainer & Docker C/R and we want it to be available
before 1.6'th release in June.

Plus we included cgroup yard destruction and properties
restore that were spotted soon after 1.5.

Signed-off-by: Pavel Emelyanov <xemul@parallels.com>