img: Remove empty lazy images after dump
If dump doesn't generate data into some image file the respective file
is not created at all as it used to be in 1.5. This brings potential
problem -- if the file with the image name exists during dump (e.g. a
dump file left from previous, maybe unsuccessful, dump) then restore
would pick this file and read bad/wrong/fake data from it.

To return the proper behavior back the file with the name of empty image
should be removed.

Signed-off-by: Pavel Emelyanov <xemul@parallels.com>