criu: Version 1.6
So, we've got quite a lot of new stuff this release :) The biggest
code-merge was POWER arch port. Another big thing that has happened
was criu integration with libcontainer. And a tiny change that will
become big next release -- the "criu show" command is going to be
thrown away in 3 monthes, the CRIT tool should be used instead.

Signed-off-by: Pavel Emelyanov <xemul@parallels.com>