aio: Don't consider aio_nr_req as a fd
When freeing the vma entries, don't call close on vm_file_fd when dealing
with a VMA AIO entry since the vm_file_fd is then filled with aio_nr_req as
part of the union.

I hit this issue when running the test aio00 on ppc64. Here the value of
the VMA aio aio_nr_req field was matching the value of the service file
descriptor IMG_FD_OFF. This leads to an obscure checkpoint error.

Signed-off-by: Laurent Dufour <ldufour@linux.vnet.ibm.com>
Signed-off-by: Pavel Emelyanov <xemul@parallels.com>