cr-dump: wait killed processes
The kill syscall queues a signal, but doesn't wait when it will be
handled.

We need to wait processes if we kill them. The user doesn't
expect to find processes after dump in this case.

PTRACE_DETACH returns errors for dead tasks, so we don't need to do it
in these cases.

Cc: Nikita Spiridonov <nspiridonov@odin.com>
Signed-off-by: Andrey Vagin <avagin@openvz.org>
Signed-off-by: Pavel Emelyanov <xemul@parallels.com>