files: Adjust file position for /dev/kmsg
The kernel doesnt allow to call lseek on /dev/kmsg
if it has been opened with O_WRONLY mode so our
restore procedure fails.

Thus if we meet a file which fits the condition
above -- set it's position to predefined value
which tells criu to not call lseek on restore.

Signed-off-by: Cyrill Gorcunov <gorcunov@openvz.org>
Signed-off-by: Pavel Emelyanov <xemul@parallels.com>