criu: Version 1.7
We have quite a lot of new stuff this time. More flexibility
in cgroup management code (and this story far from being over)
in the library and in the crit command line, preparations to
x66 32bit, that also helped the PPC port, freezer cgroup
filannly found its place in the code. We also had our first
CVE-s found :) and a LOT of bugs fixed.

So the plan for the 1.8 is: cgroups, secured containers and
bugfixes. x86 32bit is tempting too.

Signed-off-by: Pavel Emelyanov <xemul@parallels.com>